proposal for structure which uses the account object itself as the data container instead of a dictionary
