+++ program.java
@@ -1,12 +1,12 @@
-import java.util.Random;
+import java.util.HashMap;
 import java.util.Scanner;
 
 public class program 
 {
     public static void main(String[] args)
     {
+        HashMap<Integer, Account> accounts = new HashMap<>();
         Scanner scan = new Scanner(System.in);
-        Random rand = new Random();
         int choice;
 
         do 
@@ -45,12 +45,8 @@ public class program
                     break;
 
                 case 6:
-                    System.out.println("Make an Account");
-                    System.out.println("Enter your name: ");
-                    String name = scan.nextLine();
-                    int routing_number = rand.nextInt(10000001);
-                    Account account = new Account(name, routing_number, 0.00, 0.00);
-                    account.display();
+                    Account new_account = new Account();
+                    accounts.put(new_account.get_routing_number(), new_account);
                     break;
 
                 case 7:
